Adds thoughtExpiry-rules
